pyproject.toml
[project]
name = "doctolib-watcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.6",
    "brotli>=1.1.0",
    "python-dotenv>=1.1.0",
    "twilio>=9.6.2",
    "zstandard>=0.23.0",
]
